Move dashboard into new sub-app structure
This moves the dashboard concerns into their own express sub-app.
