+++ src/apps/dashboard/router.js
@@ -1,0 +1,9 @@
+const router = require('express').Router()
+
+const dashboardController = require('./dashboard.controller')
+
+router.get('/', dashboardController.getHandler)
+
+module.exports = {
+  router,
+}
